requirements.txt
asgiref==3.2.10
astroid==2.4.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
bs4==0.0.1
cachelib==0.1
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
Click==7.0
Django==3.0.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Session==0.3.2
idna==2.9
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Js2Py==0.70
lazy-object-proxy==1.4.3
Markdown==3.2.2
markdown2==2.3.9
MarkupSafe==1.1.1
mccabe==0.6.1
peewee==3.11.2
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycparser==2.19
pyjsparser==2.7.1
pylint==2.5.2
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.13.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==16.7.7
Werkzeug==1.0.1
wrapt==1.12.1
xmljson==0.2.1
xmltodict==0.12.0
zipp==3.1.0
